+++ app/Http/Controllers/KhuyenMaiSanPhamController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Validator;
 
 class KhuyenMaiSanPhamController extends Controller
 {
@@ -31,50 +32,61 @@ class KhuyenMaiSanPhamController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function index($query)
+    public function index()
     {
-//        $this->base->index($query);
-//        return response()->json($this->base->getMessage(), $this->base->getStatus());
-
-        $objs = null;
-        $code = null;
-        switch ($query) {
-            case "all":
-                $objs = DB::table(self::table)
-                    ->join(SanPhamController::table, self::table . '.' . self::ma_san_pham, '=', SanPhamController::table . '.' . SanPhamController::id)
-                    ->join(LoaiSanPhamController::table, self::table . '.' . self::ma_loai_san_pham, '=', LoaiSanPhamController::table . '.' . LoaiSanPhamController::id)
-                    ->join(ThuongHieuController::table, self::table . '.' . self::ma_thuong_hieu, '=', ThuongHieuController::table . '.' . ThuongHieuController::id)
-                    ->join(NgayKhuyenMaiController::table, self::table . '.' . self::ma_ngay_khuyen_mai, '=', NgayKhuyenMaiController::table . '.' . NgayKhuyenMaiController::id)
-                    ->select(self::table . '.*', SanPhamController::table . '.' . SanPhamController::ten_san_pham, LoaiSanPhamController::table . '.' . LoaiSanPhamController::ten_loai_san_pham, ThuongHieuController::table . '.' . ThuongHieuController::ten_thuong_hieu, NgayKhuyenMaiController::table . '.' . NgayKhuyenMaiController::ngay_gio)
-                    ->get();
-                $code = 200;
-                break;
-            case "active":
-                $objs = DB::table(self::table)
-                    ->join(SanPhamController::table, self::table . '.' . self::ma_san_pham, '=', SanPhamController::table . '.' . SanPhamController::id)
-                    ->join(LoaiSanPhamController::table, self::table . '.' . self::ma_loai_san_pham, '=', LoaiSanPhamController::table . '.' . LoaiSanPhamController::id)
-                    ->join(ThuongHieuController::table, self::table . '.' . self::ma_thuong_hieu, '=', ThuongHieuController::table . '.' . ThuongHieuController::id)
-                    ->join(NgayKhuyenMaiController::table, self::table . '.' . self::ma_ngay_khuyen_mai, '=', NgayKhuyenMaiController::table . '.' . NgayKhuyenMaiController::id)
-                    ->select(self::table . '.*', SanPhamController::table . '.' . SanPhamController::ten_san_pham, LoaiSanPhamController::table . '.' . LoaiSanPhamController::ten_loai_san_pham, ThuongHieuController::table . '.' . ThuongHieuController::ten_thuong_hieu, NgayKhuyenMaiController::table . '.' . NgayKhuyenMaiController::ngay_gio)
-                    ->where(self::table . '.' . self::isActive, '=', true)->get();
-                $code = 200;
-                break;
-            case "inactive":
-                $objs = DB::table(self::table)
-                    ->join(SanPhamController::table, self::table . '.' . self::ma_san_pham, '=', SanPhamController::table . '.' . SanPhamController::id)
-                    ->join(LoaiSanPhamController::table, self::table . '.' . self::ma_loai_san_pham, '=', LoaiSanPhamController::table . '.' . LoaiSanPhamController::id)
-                    ->join(ThuongHieuController::table, self::table . '.' . self::ma_thuong_hieu, '=', ThuongHieuController::table . '.' . ThuongHieuController::id)
-                    ->join(NgayKhuyenMaiController::table, self::table . '.' . self::ma_ngay_khuyen_mai, '=', NgayKhuyenMaiController::table . '.' . NgayKhuyenMaiController::id)
-                    ->select(self::table . '.*', SanPhamController::table . '.' . SanPhamController::ten_san_pham, LoaiSanPhamController::table . '.' . LoaiSanPhamController::ten_loai_san_pham, ThuongHieuController::table . '.' . ThuongHieuController::ten_thuong_hieu, NgayKhuyenMaiController::table . '.' . NgayKhuyenMaiController::ngay_gio)
-                    ->where(self::table . '.' . self::isActive, '=', false)->get();
-                $code = 200;
-                break;
-            default:
-                $objs = "Không tìm thấy";
-                $code = 200;
-                break;
+        $user = auth()->user();
+        $loai_tk = $user->loai_tai_khoan;
+        if ($loai_tk == TaiKhoanController::NV || $loai_tk == TaiKhoanController::QT) {
+            $objs = null;
+            $code = null;
+            $objs = DB::table(self::table)
+                ->join(SanPhamController::table, self::table . '.' . self::ma_san_pham, '=', SanPhamController::table . '.' . SanPhamController::id)
+                ->join(LoaiSanPhamController::table, self::table . '.' . self::ma_loai_san_pham, '=', LoaiSanPhamController::table . '.' . LoaiSanPhamController::id)
+                ->join(ThuongHieuController::table, self::table . '.' . self::ma_thuong_hieu, '=', ThuongHieuController::table . '.' . ThuongHieuController::id)
+                ->join(NgayKhuyenMaiController::table, self::table . '.' . self::ma_ngay_khuyen_mai, '=', NgayKhuyenMaiController::table . '.' . NgayKhuyenMaiController::id)
+                ->select(self::table . '.*', SanPhamController::table . '.' . SanPhamController::ten_san_pham, LoaiSanPhamController::table . '.' . LoaiSanPhamController::ten_loai_san_pham, ThuongHieuController::table . '.' . ThuongHieuController::ten_thuong_hieu, NgayKhuyenMaiController::table . '.' . NgayKhuyenMaiController::ngay_gio)
+                ->get();
+            $code = 200;
+//        switch ($query) {
+//            case "all":
+//                $objs = DB::table(self::table)
+//                    ->join(SanPhamController::table, self::table . '.' . self::ma_san_pham, '=', SanPhamController::table . '.' . SanPhamController::id)
+//                    ->join(LoaiSanPhamController::table, self::table . '.' . self::ma_loai_san_pham, '=', LoaiSanPhamController::table . '.' . LoaiSanPhamController::id)
+//                    ->join(ThuongHieuController::table, self::table . '.' . self::ma_thuong_hieu, '=', ThuongHieuController::table . '.' . ThuongHieuController::id)
+//                    ->join(NgayKhuyenMaiController::table, self::table . '.' . self::ma_ngay_khuyen_mai, '=', NgayKhuyenMaiController::table . '.' . NgayKhuyenMaiController::id)
+//                    ->select(self::table . '.*', SanPhamController::table . '.' . SanPhamController::ten_san_pham, LoaiSanPhamController::table . '.' . LoaiSanPhamController::ten_loai_san_pham, ThuongHieuController::table . '.' . ThuongHieuController::ten_thuong_hieu, NgayKhuyenMaiController::table . '.' . NgayKhuyenMaiController::ngay_gio)
+//                    ->get();
+//                $code = 200;
+//                break;
+//            case "active":
+//                $objs = DB::table(self::table)
+//                    ->join(SanPhamController::table, self::table . '.' . self::ma_san_pham, '=', SanPhamController::table . '.' . SanPhamController::id)
+//                    ->join(LoaiSanPhamController::table, self::table . '.' . self::ma_loai_san_pham, '=', LoaiSanPhamController::table . '.' . LoaiSanPhamController::id)
+//                    ->join(ThuongHieuController::table, self::table . '.' . self::ma_thuong_hieu, '=', ThuongHieuController::table . '.' . ThuongHieuController::id)
+//                    ->join(NgayKhuyenMaiController::table, self::table . '.' . self::ma_ngay_khuyen_mai, '=', NgayKhuyenMaiController::table . '.' . NgayKhuyenMaiController::id)
+//                    ->select(self::table . '.*', SanPhamController::table . '.' . SanPhamController::ten_san_pham, LoaiSanPhamController::table . '.' . LoaiSanPhamController::ten_loai_san_pham, ThuongHieuController::table . '.' . ThuongHieuController::ten_thuong_hieu, NgayKhuyenMaiController::table . '.' . NgayKhuyenMaiController::ngay_gio)
+//                    ->where(self::table . '.' . self::isActive, '=', true)->get();
+//                $code = 200;
+//                break;
+//            case "inactive":
+//                $objs = DB::table(self::table)
+//                    ->join(SanPhamController::table, self::table . '.' . self::ma_san_pham, '=', SanPhamController::table . '.' . SanPhamController::id)
+//                    ->join(LoaiSanPhamController::table, self::table . '.' . self::ma_loai_san_pham, '=', LoaiSanPhamController::table . '.' . LoaiSanPhamController::id)
+//                    ->join(ThuongHieuController::table, self::table . '.' . self::ma_thuong_hieu, '=', ThuongHieuController::table . '.' . ThuongHieuController::id)
+//                    ->join(NgayKhuyenMaiController::table, self::table . '.' . self::ma_ngay_khuyen_mai, '=', NgayKhuyenMaiController::table . '.' . NgayKhuyenMaiController::id)
+//                    ->select(self::table . '.*', SanPhamController::table . '.' . SanPhamController::ten_san_pham, LoaiSanPhamController::table . '.' . LoaiSanPhamController::ten_loai_san_pham, ThuongHieuController::table . '.' . ThuongHieuController::ten_thuong_hieu, NgayKhuyenMaiController::table . '.' . NgayKhuyenMaiController::ngay_gio)
+//                    ->where(self::table . '.' . self::isActive, '=', false)->get();
+//                $code = 200;
+//                break;
+//            default:
+//                $objs = "Không tìm thấy";
+//                $code = 200;
+//                break;
+//        }
+            return response()->json($objs, $code);
+        } else {
+            return response()->json('Tài khoản không đủ quyền truy cập', 200);
         }
-        return response()->json($objs, $code);
     }
 
     /**
@@ -96,13 +108,22 @@ class KhuyenMaiSanPhamController extends Controller
      */
     public function store(Request $request)
     {
-        $this->validate($request, [
-            self::muc_khuyen_mai => 'required',
-            self::ma_ngay_khuyen_mai => 'required',
-        ]);
+        $user = auth()->user();
+        $loai_tk = $user->loai_tai_khoan;
+        if ($loai_tk == TaiKhoanController::NV || $loai_tk == TaiKhoanController::QT) {
+            $validator = Validator::make($request->all(), [
+                self::muc_khuyen_mai => 'required',
+                self::ma_ngay_khuyen_mai => 'required',
+            ]);
+            if ($validator->fails()) {
+                return response()->json(['error' => $validator->errors()->all()], 200);
+            }
 
-        $this->base->store($request);
-        return response()->json($this->base->getMessage(), $this->base->getStatus());
+            $this->base->store($request);
+            return response()->json($this->base->getMessage(), $this->base->getStatus());
+        } else {
+            return response()->json('Tài khoản không đủ quyền truy cập', 200);
+        }
     }
 
     /**
@@ -113,8 +134,14 @@ class KhuyenMaiSanPhamController extends Controller
      */
     public function show($id)
     {
-        $this->base->show($id);
-        return response()->json($this->base->getMessage(), $this->base->getStatus());
+        $user = auth()->user();
+        $loai_tk = $user->loai_tai_khoan;
+        if ($loai_tk == TaiKhoanController::NV || $loai_tk == TaiKhoanController::QT) {
+            $this->base->show($id);
+            return response()->json($this->base->getMessage(), $this->base->getStatus());
+        } else {
+            return response()->json('Tài khoản không đủ quyền truy cập', 200);
+        }
     }
 
     /**
@@ -137,8 +164,14 @@ class KhuyenMaiSanPhamController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $this->base->update($request, $id);
-        return response()->json($this->base->getMessage(), $this->base->getStatus());
+        $user = auth()->user();
+        $loai_tk = $user->loai_tai_khoan;
+        if ($loai_tk == TaiKhoanController::NV || $loai_tk == TaiKhoanController::QT) {
+            $this->base->update($request, $id);
+            return response()->json($this->base->getMessage(), $this->base->getStatus());
+        } else {
+            return response()->json('Tài khoản không đủ quyền truy cập', 200);
+        }
     }
 
     /**
@@ -149,7 +182,13 @@ class KhuyenMaiSanPhamController extends Controller
      */
     public function destroy(Request $request)
     {
-        $this->base->destroy($request);
-        return response()->json($this->base->getMessage(), $this->base->getStatus());
+        $user = auth()->user();
+        $loai_tk = $user->loai_tai_khoan;
+        if ($loai_tk == TaiKhoanController::NV || $loai_tk == TaiKhoanController::QT) {
+            $this->base->destroy($request);
+            return response()->json($this->base->getMessage(), $this->base->getStatus());
+        } else {
+            return response()->json('Tài khoản không đủ quyền truy cập', 200);
+        }
     }
 }
