+++ app/Http/Controllers/NgayKhuyenMaiController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Validator;
 
 class NgayKhuyenMaiController extends Controller
 {
@@ -20,15 +21,22 @@ class NgayKhuyenMaiController extends Controller
     {
         $this->base = new BaseController(self::table, self::id, self::isActive);
     }
+
     /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
      */
-    public function index($query)
+    public function index()
     {
-        $this->base->index($query);
-        return response()->json($this->base->getMessage(), $this->base->getStatus());
+        $user = auth()->user();
+        $loai_tk = $user->loai_tai_khoan;
+        if ($loai_tk == TaiKhoanController::NV || $loai_tk == TaiKhoanController::QT) {
+            $this->base->index();
+            return response()->json($this->base->getMessage(), $this->base->getStatus());
+        } else {
+            return response()->json('Tài khoản không đủ quyền truy cập', 200);
+        }
     }
 
     /**
@@ -50,12 +58,21 @@ class NgayKhuyenMaiController extends Controller
      */
     public function store(Request $request)
     {
-        $this->validate($request, [
-            self::ngay_gio => 'required',
-        ]);
+        $user = auth()->user();
+        $loai_tk = $user->loai_tai_khoan;
+        if ($loai_tk == TaiKhoanController::NV || $loai_tk == TaiKhoanController::QT) {
+            $validator = Validator::make($request->all(), [
+                self::ngay_gio => 'required',
+            ]);
+            if ($validator->fails()) {
+                return response()->json(['error' => $validator->errors()->all()], 200);
+            }
 
-        $this->base->store($request);
-        return response()->json($this->base->getMessage(), $this->base->getStatus());
+            $this->base->store($request);
+            return response()->json($this->base->getMessage(), $this->base->getStatus());
+        } else {
+            return response()->json('Tài khoản không đủ quyền truy cập', 200);
+        }
     }
 
     /**
@@ -66,8 +83,14 @@ class NgayKhuyenMaiController extends Controller
      */
     public function show($id)
     {
-        $this->base->show($id);
-        return response()->json($this->base->getMessage(), $this->base->getStatus());
+        $user = auth()->user();
+        $loai_tk = $user->loai_tai_khoan;
+        if ($loai_tk == TaiKhoanController::NV || $loai_tk == TaiKhoanController::QT) {
+            $this->base->show($id);
+            return response()->json($this->base->getMessage(), $this->base->getStatus());
+        } else {
+            return response()->json('Tài khoản không đủ quyền truy cập', 200);
+        }
     }
 
     /**
@@ -90,8 +113,14 @@ class NgayKhuyenMaiController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $this->base->update($request, $id);
-        return response()->json($this->base->getMessage(), $this->base->getStatus());
+        $user = auth()->user();
+        $loai_tk = $user->loai_tai_khoan;
+        if ($loai_tk == TaiKhoanController::NV || $loai_tk == TaiKhoanController::QT) {
+            $this->base->update($request, $id);
+            return response()->json($this->base->getMessage(), $this->base->getStatus());
+        } else {
+            return response()->json('Tài khoản không đủ quyền truy cập', 200);
+        }
     }
 
     /**
@@ -102,7 +131,13 @@ class NgayKhuyenMaiController extends Controller
      */
     public function destroy(Request $request)
     {
-        $this->base->destroy($request);
-        return response()->json($this->base->getMessage(), $this->base->getStatus());
+        $user = auth()->user();
+        $loai_tk = $user->loai_tai_khoan;
+        if ($loai_tk == TaiKhoanController::NV || $loai_tk == TaiKhoanController::QT) {
+            $this->base->destroy($request);
+            return response()->json($this->base->getMessage(), $this->base->getStatus());
+        } else {
+            return response()->json('Tài khoản không đủ quyền truy cập', 200);
+        }
     }
 }
