+++ app/Http/Controllers/NhaCungCapController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Validator;
 
 class NhaCungCapController extends Controller
 {
@@ -31,10 +32,16 @@ class NhaCungCapController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function index($query)
+    public function index()
     {
-        $this->base->index($query);
-        return response()->json($this->base->getMessage(), $this->base->getStatus());
+        $user = auth()->user();
+        $loai_tk = $user->loai_tai_khoan;
+        if ($loai_tk == TaiKhoanController::NV || $loai_tk == TaiKhoanController::QT) {
+            $this->base->index();
+            return response()->json($this->base->getMessage(), $this->base->getStatus());
+        } else {
+            return response()->json('Tài khoản không đủ quyền truy cập', 200);
+        }
     }
 
     /**
@@ -54,14 +61,23 @@ class NhaCungCapController extends Controller
      */
     public function store(Request $request)
     {
-        $this->validate($request, [
-            self::ten => 'required',
-            self::dia_chi => 'required',
-            self::so_dien_thoai => 'required',
-        ]);
+        $user = auth()->user();
+        $loai_tk = $user->loai_tai_khoan;
+        if ($loai_tk == TaiKhoanController::NV || $loai_tk == TaiKhoanController::QT) {
+            $validator = Validator::make($request->all(), [
+                self::ten => 'required',
+                self::dia_chi => 'required',
+                self::so_dien_thoai => 'required',
+            ]);
+            if ($validator->fails()) {
+                return response()->json(['error' => $validator->errors()->all()], 200);
+            }
 
-        $this->base->store($request);
-        return response()->json($this->base->getMessage(), $this->base->getStatus());
+            $this->base->store($request);
+            return response()->json($this->base->getMessage(), $this->base->getStatus());
+        } else {
+            return response()->json('Tài khoản không đủ quyền truy cập', 200);
+        }
     }
 
     /**
@@ -72,8 +88,14 @@ class NhaCungCapController extends Controller
      */
     public function show($id)
     {
-        $this->base->show($id);
-        return response()->json($this->base->getMessage(), $this->base->getStatus());
+        $user = auth()->user();
+        $loai_tk = $user->loai_tai_khoan;
+        if ($loai_tk == TaiKhoanController::NV || $loai_tk == TaiKhoanController::QT) {
+            $this->base->show($id);
+            return response()->json($this->base->getMessage(), $this->base->getStatus());
+        } else {
+            return response()->json('Tài khoản không đủ quyền truy cập', 200);
+        }
     }
 
     /**
@@ -96,8 +118,14 @@ class NhaCungCapController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $this->base->update($request, $id);
-        return response()->json($this->base->getMessage(), $this->base->getStatus());
+        $user = auth()->user();
+        $loai_tk = $user->loai_tai_khoan;
+        if ($loai_tk == TaiKhoanController::NV || $loai_tk == TaiKhoanController::QT) {
+            $this->base->update($request, $id);
+            return response()->json($this->base->getMessage(), $this->base->getStatus());
+        } else {
+            return response()->json('Tài khoản không đủ quyền truy cập', 200);
+        }
     }
 
     /**
@@ -108,7 +136,13 @@ class NhaCungCapController extends Controller
      */
     public function destroy(Request $request)
     {
-        $this->base->destroy($request);
-        return response()->json($this->base->getMessage(), $this->base->getStatus());
+        $user = auth()->user();
+        $loai_tk = $user->loai_tai_khoan;
+        if ($loai_tk == TaiKhoanController::NV || $loai_tk == TaiKhoanController::QT) {
+            $this->base->destroy($request);
+            return response()->json($this->base->getMessage(), $this->base->getStatus());
+        } else {
+            return response()->json('Tài khoản không đủ quyền truy cập', 200);
+        }
     }
 }
